Decision Support Agent/requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
tiktoken
python-dotenv

yfinance
pandas
numpy

plotly
kaleido

astunparse
httpx